+++ dplypy/test/write_file_test.py
@@ -1,0 +1,63 @@
+import os
+
+import pandas as pd
+import pytest
+
+from dplypy.dplypy import DplyFrame, write_file
+
+def test_write_file():
+    pandas_df = pd.DataFrame(
+        data={
+            "col1": [0, 1, 2, 3],
+            "col2": [3, 4, 5, 6],
+            "col3": [6, 7, 8, 9],
+            "col4": [9, 10, 11, 12],
+        }
+    )
+
+    df = DplyFrame(pandas_df)
+
+    # To csv file
+    (
+        df
+        + write_file("df_no_index.csv", sep=",", index=False)
+        + write_file("df_with_index.csv", sep=",", index=True)
+    )
+    read_df = pd.read_csv("df_no_index.csv", sep=",")
+    pd.testing.assert_frame_equal(df.pandas_df, read_df)  # Without index
+    os.remove("df_no_index.csv")
+
+    read_df = pd.read_csv("df_with_index.csv", sep=",", index_col=0)
+    pd.testing.assert_frame_equal(df.pandas_df, read_df)  # With index
+    os.remove("df_with_index.csv")
+
+    # To excel file
+    # Requires dependency openpyxl
+    (
+        df
+        + write_file("df_no_index.xlsx", index=False)
+        + write_file("df_with_index.xlsx", index=True)
+    )
+    read_df = pd.read_excel("df_no_index.xlsx", engine="openpyxl")
+    pd.testing.assert_frame_equal(df.pandas_df, read_df)  # Without index
+    os.remove("df_no_index.xlsx")
+
+    read_df = pd.read_excel("df_with_index.xlsx", index_col=0, engine="openpyxl")
+    pd.testing.assert_frame_equal(df.pandas_df, read_df)  # Without index
+    os.remove("df_with_index.xlsx")
+
+    # To json
+    df + write_file("df_no_index.json")
+    read_df = pd.read_json("df_no_index.json")
+    pd.testing.assert_frame_equal(df.pandas_df, read_df)  # Without index
+    os.remove("df_no_index.json")
+
+    # To pickle
+    df + write_file("df_no_index.pkl")
+    read_df = pd.read_pickle("df_no_index.pkl")
+    pd.testing.assert_frame_equal(df.pandas_df, read_df)  # Without index
+    os.remove("df_no_index.pkl")
+
+    # Error case
+    with pytest.raises(IOError):
+        df + write_file("df.abc")
